Add basic support for pushing blob files to typesense
It listens to blob events, translates the metadata if needed, merges
the metadata with the person info, and then pushes the resulting
documents to typesense in a worker.

Also add a CLI command for deleting blob files, for easier testing.
And a filename option to the upload command to test with different filenames.

With the upload command, and the blob integration, test files can now be added
via an external script, so remove the code for adding dummy file typesense entries.