Implement basic support for syncing in blob files
After the person sync is done we fetch only the base data for all
persons (to reduce memory usage) for a mappig of obfuscated IDs to
base info.

After that we get blob files in chunks, translate them to typesense
and add the missing base data, and finally add them to typesense.