typesense: Update the base data for all related DocumentFile entries if a Person entry changes
This happens for partial syncs, and for single syncs. This results in a request to typesense
per person, so this will slow down those operations, or limit them. We might need to
fall back to a full sync, or get the full person->base mapping if there are too many changes
in the future.