blob sync: handle the document translation in the async job
To avoid dealing with multiple ways to translate a fileData object
to typesense just pass the blob ID to the worker and fetch the
fileData there vie the API. This way the code doing the initial indexing
and the updates can be shared.

This also means less blocking in the blob event handler.