sync: store a schema version in the collection metadata
When creating a new collection we store a schema version string in
the collection metadata, which is provided by the translator/user.

In case the version of the collection in typesense doesn't match
the current version we force a full sync. This allows us to automatically
update the collection if either the typesense schema, or the translator
code changes.