typesense: fix creation of an empty collection when no alias exists
When no alias exists initially we try to set up the DB with a minimal working
database, so accessinf the alias works at least. This was recently broken
by passing an empty schema for the dummy collectin, which isn't allowed
by typesense. Fix this by creating a dummy schema which auto detects all
fields, since that is non-specific and the collection will always be empty
anyway.

This fixes the initial typesense DB setup in case it is empty.