PersonSyncResultInterface: add method for detecting if the result is a full sync
Even in case the connector is asked for a partial sync, it might, for performance
reasons start a full sync internally, for example if there are too many changes
since the last sync, and single change updates are slow.

By communicating this to the cabinet bundle, the cabinet bundle can decide to
rebuild its database as well, even if it just asked for a partial sync.