typesense: do a full sync if the connector tells us to
In case the connector returns that it did a full sync even though
we passed a cursor, use that information to start a full sync on the
cabinet end as well.

This allows the connector to fall back to a full sync in case there are
too many changes, and the cabinet bundle to use that information to
optimize for that case.