typesense: create a scoped read-only API key for the proxy
We don't want users that are allowed to access the API to have access to
other collections in typesense and to have write access on the cabinet collection.

To fix this create a proxy key during setup (the key value doesn't matter really,
since it is only used internally) and use that key when talking to typesense when
called via the proxy endpoint.