DocumentTranslationEvent: extend to allow translating to multiple documents
This allows the translation handler to either set to an empty array, to
skip one document all together, or set multiple documents, so the result
is multiple typesense documents per input.

This currently assumes that documents only get added and not removed,
since there is no logic to delete additional documents right now. This
might need to be fixed in the future.