TypesenseSync: add DocumentFinalizeEvent
This event is called after all the transformation and merging is done and
right before the document is sent to typesense.

This allows the user to reference keys between the person and file document
and update some fields depending on that if needed.

In our case we need to look up study IDs from the person to get a display
name for showing in some file documents that only contain a sutdy ID.