typesense proxy: enable longer server side caching
Enable one hour of caching in typesense.

To set the cache TTL we have to use a scoped key, and scoped
keys are only possible for search-only keys, so we now create
two keys, one for searching and one for everything else, and
use then create a scoped based on the first one.

Since typesense doesn't do any caching invalidation we have to do
it manually by calling the cache/clear endpoint.

It's not clear yet how many cache hits we get with this, and/or
if we should cache in the proxy instead, but let's give it a try.

The scoped key thing might also be useful in the future for other
things.