Add a partitioned search feature
Allows the bundle to split up a request into multiple sub requests
that work on a subset of the dataset and at the end meges all the responses
again to what a typesense client might expect. The splitting is based
on a special field that contains 0-99, and it's assumed that all documents
from the same group have the same number, so they always are grouped into
the same input set (the special field is hardcoded atm, person.partitionKey)

This has some downsides in that we can't do pagination properly, that
total facet counts are no longer exact, etc. But it allows to speed up
the search assuming there are enough cores on the server.

There is some room for improvements here still, like proxying pagination,
implement more of the result sort syntax, implement the facet sort syntax
etc. but the basics work.

By default this is disabled.