typesense: add a isPrimary field (experiment)
This field identifies the primary document of a group. In case the transformer
returns multiple documents we mark the first one as primary.

This could in theory be used to get one entry per group without doing
group_by, as long as there are no fields queried that are different
between the group members.

As such this only makes sense if we only facet on group specific fields
on demand, which we currently don't.