sync-person-actions: allow to sync via the typesense document ID
We currently only supported syncs via the person ID but that depends
on how the schema is configured.

To allow syncing without knowing the schema we allow passing in a
typesense document ID instead. In the backend we translate that
to a person ID and do the same things as before.

Keep the old query param for not to make the transition easier.