sync: add support for isFullSyncResult()
In case the connector reports to us that a full sync happened
we want to use that to create a new collection instead of updating
all existing records.

Also rename "full" to "forceFull", since a full sync can happen
anyway in case the schema is outdated, or the backend tells us,
so full=false doesn't mean it could lead to a full sync.