sync: cron; sync config; locking
Adds two new bundle config settings:

sync.schedule: the cron schedule for doing syncs
sync.full_sync_interval: when to force a full sync

Adds locking to all typesense access. All sync operations and read
operations are serialized via one lock atm. This has the downside
that on a full sync all operations are paused, so this likely needs
to be improved in the future somehow (splitting the sync into two steps
so we can detect early if we'll do a full sync)

We also add a lock for syncing, so that only one sync is queued at any time,
to avoid them piling up (also they take a lot of resources).

To test locking with setup this adds a ":purge" command which deletes all
typesense data.