partition: add partial support for drop_tokens_threshold and typo_tokens_threshold
drop_tokens_threshold and typo_tokens_threshold only kick in in typesense once there
are not enough results returned. Since we deal with queries on a subset of the available
data the result might be empty, but the overall result count might be enough to not
need them.

To avoid for example typo tolerance kickin in in case there are enough results overall
we disable them, then merge the results, and in case there are not enough resutls
we retry the request with typo_tokens_threshold restored.

This way the user at least gets either exact matches, or matches with typo tolerance,
even if the results might be different as "typo_tokens_threshold" is counted for each
sub collection instead of the whole collection.

Retrying also shouldn't be a perforamnce issue here, as searches that don't return any
results are cheap.